Fix cURL library API usage bug: the library expects a long, not an int.

+++ indra/llcorehttp/_httplibcurl.cpp
@@ -333,8 +333,11 @@ bool HttpLibcurl::completeRequest(CURLM * multi_handle, CURL * handle, CURLcode 
 	}
 	if (op->mStatus)
 	{
-		int http_status(HTTP_OK);
-
+		//<FS:TS> The cURL library doc specifies that you need to pass
+		// a pointer to a long, not a pointer to an int, here.
+		//int http_status(HTTP_OK);
+		long int http_status(HTTP_OK);
+                                
 		curl_easy_getinfo(handle, CURLINFO_RESPONSE_CODE, &http_status);
 		if (http_status >= 100 && http_status <= 999)
 		{
